+++ src/acceptance-test/java/org/zalando/nakadi/repository/kafka/KafkaRepositoryAT.java
@@ -6,14 +6,15 @@ import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.common.PartitionInfo;
 import org.echocat.jomon.runtime.concurrent.RetryForSpecifiedTimeStrategy;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.zalando.nakadi.config.NakadiSettings;
 import org.zalando.nakadi.domain.BatchFactory;
 import org.zalando.nakadi.domain.BatchItem;
+import org.zalando.nakadi.domain.CleanupPolicy;
 import org.zalando.nakadi.domain.EventPublishingStatus;
+import org.zalando.nakadi.repository.NakadiTopicConfig;
 import org.zalando.nakadi.repository.zookeeper.ZooKeeperHolder;
 import org.zalando.nakadi.repository.zookeeper.ZookeeperSettings;
 import org.zalando.nakadi.util.UUIDGenerator;
@@ -23,6 +24,7 @@ import org.zalando.nakadi.webservice.BaseAT;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.UUID;
 
 import static org.echocat.jomon.runtime.concurrent.Retryer.executeWithRetry;
@@ -42,13 +44,17 @@ public class KafkaRepositoryAT extends BaseAT {
     private static final int DEFAULT_REPLICA_FACTOR = 1;
     private static final int MAX_TOPIC_PARTITION_COUNT = 10;
     private static final int DEFAULT_TOPIC_ROTATION = 50000000;
+    private static final int COMPACTED_TOPIC_ROTATION = 60000;
+    private static final int COMPACTED_TOPIC_SEGMENT_BYTES = 1000000;
+    private static final int COMPACTED_TOPIC_COMPACTION_LAG = 1000;
     private static final int DEFAULT_COMMIT_TIMEOUT = 60;
     private static final int ZK_SESSION_TIMEOUT = 30000;
     private static final int ZK_CONNECTION_TIMEOUT = 10000;
     private static final int NAKADI_SEND_TIMEOUT = 10000;
     private static final int NAKADI_POLL_TIMEOUT = 10000;
-    private static final Long RETENTION_TIME = 100L;
+    private static final Long DEFAULT_RETENTION_TIME = 100L;
     private static final Long DEFAULT_TOPIC_RETENTION = 100000000L;
+    private static final CleanupPolicy DEFAULT_CLEANUP_POLICY = CleanupPolicy.DELETE;
     private static final int KAFKA_REQUEST_TIMEOUT = 30000;
     private static final int KAFKA_BATCH_SIZE = 1048576;
     private static final long KAFKA_LINGER_MS = 0;
@@ -66,6 +72,8 @@ public class KafkaRepositoryAT extends BaseAT {
     private ZookeeperSettings zookeeperSettings;
     private KafkaTestHelper kafkaHelper;
     private KafkaTopicRepository kafkaTopicRepository;
+    private NakadiTopicConfig defaultTopicConfig;
+    private KafkaTopicConfigFactory kafkaTopicConfigFactory;
 
     @Before
     public void setup() {
@@ -88,14 +96,19 @@ public class KafkaRepositoryAT extends BaseAT {
                 KAFKA_LINGER_MS, KAFKA_ENABLE_AUTO_COMMIT, KAFKA_MAX_REQUEST_SIZE);
         zookeeperSettings = new ZookeeperSettings(ZK_SESSION_TIMEOUT, ZK_CONNECTION_TIMEOUT);
         kafkaHelper = new KafkaTestHelper(KAFKA_URL);
+        defaultTopicConfig = new NakadiTopicConfig(DEFAULT_PARTITION_COUNT, DEFAULT_CLEANUP_POLICY,
+                Optional.of(DEFAULT_RETENTION_TIME));
+        kafkaTopicConfigFactory = new KafkaTopicConfigFactory(new UUIDGenerator(), DEFAULT_REPLICA_FACTOR,
+                DEFAULT_TOPIC_ROTATION, COMPACTED_TOPIC_ROTATION, COMPACTED_TOPIC_SEGMENT_BYTES,
+                COMPACTED_TOPIC_COMPACTION_LAG);
         kafkaTopicRepository = createKafkaTopicRepository();
     }
 
     @Test(timeout = 10000)
     @SuppressWarnings("unchecked")
-    public void whenCreateTopicThenTopicIsCreated() throws Exception {
+    public void whenCreateTopicThenTopicIsCreated() {
         // ACT //
-        final String topicName = kafkaTopicRepository.createTopic(DEFAULT_PARTITION_COUNT, RETENTION_TIME);
+        final String topicName = kafkaTopicRepository.createTopic(defaultTopicConfig);
 
         // ASSERT //
         executeWithRetry(() -> {
@@ -107,6 +120,52 @@ public class KafkaRepositoryAT extends BaseAT {
 
                     partitionInfos.forEach(pInfo ->
                             assertThat(pInfo.replicas(), arrayWithSize(DEFAULT_REPLICA_FACTOR)));
+
+                    final Long retentionTime = KafkaTestHelper.getTopicRetentionTime(topicName, ZOOKEEPER_URL);
+                    assertThat(retentionTime, equalTo(DEFAULT_RETENTION_TIME));
+
+                    final String cleanupPolicy = KafkaTestHelper.getTopicCleanupPolicy(topicName, ZOOKEEPER_URL);
+                    assertThat(cleanupPolicy, equalTo("delete"));
+
+                    final String segmentMs = KafkaTestHelper.getTopicProperty(topicName, ZOOKEEPER_URL, "segment.ms");
+                    assertThat(segmentMs, equalTo(String.valueOf(DEFAULT_TOPIC_ROTATION)));
+                },
+                new RetryForSpecifiedTimeStrategy<Void>(5000).withExceptionsThatForceRetry(AssertionError.class)
+                        .withWaitBetweenEachTry(500));
+    }
+
+    @Test(timeout = 10000)
+    @SuppressWarnings("unchecked")
+    public void whenCreateCompactedTopicThenTopicIsCreated() {
+        // ACT //
+        final NakadiTopicConfig compactedTopicConfig = new NakadiTopicConfig(DEFAULT_PARTITION_COUNT,
+                CleanupPolicy.COMPACT, Optional.empty());
+        final String topicName = kafkaTopicRepository.createTopic(compactedTopicConfig);
+
+        // ASSERT //
+        executeWithRetry(() -> {
+                    final Map<String, List<PartitionInfo>> topics = getAllTopics();
+                    assertThat(topics.keySet(), hasItem(topicName));
+
+                    final List<PartitionInfo> partitionInfos = topics.get(topicName);
+                    assertThat(partitionInfos, hasSize(DEFAULT_PARTITION_COUNT));
+
+                    partitionInfos.forEach(pInfo ->
+                            assertThat(pInfo.replicas(), arrayWithSize(DEFAULT_REPLICA_FACTOR)));
+
+                    final String cleanupPolicy = KafkaTestHelper.getTopicCleanupPolicy(topicName, ZOOKEEPER_URL);
+                    assertThat(cleanupPolicy, equalTo("compact"));
+
+                    final String segmentMs = KafkaTestHelper.getTopicProperty(topicName, ZOOKEEPER_URL, "segment.ms");
+                    assertThat(segmentMs, equalTo(String.valueOf(COMPACTED_TOPIC_ROTATION)));
+
+                    final String segmentBytes = KafkaTestHelper.getTopicProperty(topicName, ZOOKEEPER_URL,
+                            "segment.bytes");
+                    assertThat(segmentBytes, equalTo(String.valueOf(COMPACTED_TOPIC_SEGMENT_BYTES)));
+
+                    final String compactionLag = KafkaTestHelper.getTopicProperty(topicName, ZOOKEEPER_URL,
+                            "min.compaction.lag.ms");
+                    assertThat(compactionLag, equalTo(String.valueOf(COMPACTED_TOPIC_COMPACTION_LAG)));
                 },
                 new RetryForSpecifiedTimeStrategy<Void>(5000).withExceptionsThatForceRetry(AssertionError.class)
                         .withWaitBetweenEachTry(500));
@@ -114,7 +173,7 @@ public class KafkaRepositoryAT extends BaseAT {
 
     @Test(timeout = 20000)
     @SuppressWarnings("unchecked")
-    public void whenDeleteTopicThenTopicIsDeleted() throws Exception {
+    public void whenDeleteTopicThenTopicIsDeleted() {
 
         // ARRANGE //
         final String topicName = UUID.randomUUID().toString();
@@ -140,7 +199,7 @@ public class KafkaRepositoryAT extends BaseAT {
     }
 
     @Test(timeout = 10000)
-    public void whenBulkSendSuccessfullyThenUpdateBatchItemStatus() throws Exception {
+    public void whenBulkSendSuccessfullyThenUpdateBatchItemStatus() {
         final List<BatchItem> items = new ArrayList<>();
         final String topicId = TestUtils.randomValidEventTypeName();
         kafkaHelper.createTopic(topicId, ZOOKEEPER_URL);
@@ -156,20 +215,6 @@ public class KafkaRepositoryAT extends BaseAT {
         for (int i = 0; i < 10; i++) {
             assertThat(items.get(i).getResponse().getPublishingStatus(), equalTo(EventPublishingStatus.SUBMITTED));
         }
-    }
-
-    @Test(timeout = 10000)
-    @SuppressWarnings("unchecked")
-    public void whenCreateTopicWithRetentionTime() throws Exception {
-        // ACT //
-        final String topicName = kafkaTopicRepository.createTopic(DEFAULT_PARTITION_COUNT, RETENTION_TIME);
-
-        // ASSERT //
-        executeWithRetry(() -> Assert.assertEquals(
-                KafkaTestHelper.getTopicRetentionTime(topicName, ZOOKEEPER_URL), RETENTION_TIME),
-                new RetryForSpecifiedTimeStrategy<Void>(5000)
-                        .withExceptionsThatForceRetry(AssertionError.class)
-                        .withWaitBetweenEachTry(500));
     }
 
     private Map<String, List<PartitionInfo>> getAllTopics() {
@@ -203,7 +248,7 @@ public class KafkaRepositoryAT extends BaseAT {
                 nakadiSettings,
                 kafkaSettings,
                 zookeeperSettings,
-                new UUIDGenerator());
+                kafkaTopicConfigFactory);
     }
 
 }