+++ src/acceptance-test/java/org/zalando/nakadi/webservice/UserJourneyAT.java
@@ -90,8 +90,9 @@ public class UserJourneyAT extends RealEnvironmentAT {
                 .body("name", equalTo(eventTypeName))
                 .body("owning_application", equalTo(owningApp))
                 .body("category", equalTo("undefined"))
-                .body("audience", equalTo("external_public"))
+                .body("audience", equalTo("external-public"))
                 .body("ordering_key_fields", equalTo(Lists.newArrayList("foo", "bar.baz")))
+                .body("cleanup_policy", equalTo("delete"))
                 .body("schema.type", equalTo("json_schema"))
                 .body("schema.schema", equalTo("{\"type\": \"object\", \"properties\": {\"foo\": " +
                         "{\"type\": \"string\"}, \"bar\": {\"type\": \"object\", \"properties\": " +