+++ src/main/java/org/zalando/nakadi/service/EventPublisher.java
@@ -9,12 +9,14 @@ import org.zalando.nakadi.config.NakadiSettings;
 import org.zalando.nakadi.domain.BatchFactory;
 import org.zalando.nakadi.domain.BatchItem;
 import org.zalando.nakadi.domain.BatchItemResponse;
+import org.zalando.nakadi.domain.CleanupPolicy;
 import org.zalando.nakadi.domain.EventPublishResult;
 import org.zalando.nakadi.domain.EventPublishingStatus;
 import org.zalando.nakadi.domain.EventPublishingStep;
 import org.zalando.nakadi.domain.EventType;
 import org.zalando.nakadi.domain.Timeline;
 import org.zalando.nakadi.enrichment.Enrichment;
+import org.zalando.nakadi.exceptions.CompactionException;
 import org.zalando.nakadi.exceptions.EnrichmentException;
 import org.zalando.nakadi.exceptions.InternalNakadiException;
 import org.zalando.nakadi.exceptions.NoSuchEventTypeException;
@@ -51,7 +53,6 @@ public class EventPublisher {
     private final Enrichment enrichment;
     private final TimelineSync timelineSync;
     private final AuthorizationValidator authValidator;
-    private final FeatureToggleService featureToggleService;
 
     @Autowired
     public EventPublisher(final TimelineService timelineService,
@@ -60,8 +61,7 @@ public class EventPublisher {
                           final Enrichment enrichment,
                           final NakadiSettings nakadiSettings,
                           final TimelineSync timelineSync,
-                          final AuthorizationValidator authValidator,
-                          final FeatureToggleService featureToggleService) {
+                          final AuthorizationValidator authValidator) {
         this.timelineService = timelineService;
         this.eventTypeCache = eventTypeCache;
         this.partitionResolver = partitionResolver;
@@ -69,7 +69,6 @@ public class EventPublisher {
         this.nakadiSettings = nakadiSettings;
         this.timelineSync = timelineSync;
         this.authValidator = authValidator;
-        this.featureToggleService = featureToggleService;
     }
 
     public EventPublishResult publish(final String events, final String eventTypeName)
@@ -99,6 +98,7 @@ public class EventPublisher {
 
             validate(batch, eventType);
             partition(batch, eventType);
+            compact(batch, eventType);
             enrich(batch, eventType);
             submit(batch, eventType);
 
@@ -108,6 +108,9 @@ public class EventPublisher {
             return aborted(EventPublishingStep.VALIDATING, batch);
         } catch (final PartitioningException e) {
             LOG.debug("Event partition error: {}", e.getMessage());
+            return aborted(EventPublishingStep.PARTITIONING, batch);
+        } catch (final CompactionException e) {
+            LOG.debug("Event compaction error: {}", e.getMessage());
             return aborted(EventPublishingStep.PARTITIONING, batch);
         } catch (final EnrichmentException e) {
             LOG.debug("Event enrichment error: {}", e.getMessage());
@@ -160,6 +163,17 @@ public class EventPublisher {
             } catch (final PartitioningException e) {
                 item.updateStatusAndDetail(EventPublishingStatus.FAILED, e.getMessage());
                 throw e;
+            }
+        }
+    }
+
+    private void compact(final List<BatchItem> batch, final EventType eventType) throws CompactionException {
+        if (eventType.getCleanupPolicy() == CleanupPolicy.COMPACT) {
+            for (final BatchItem item : batch) {
+                final String compactionKey = item.getEvent()
+                        .getJSONObject("metadata")
+                        .getString("partition_compaction_key");
+                item.setEventKey(compactionKey);
             }
         }
     }